feat: Fix provider order using react router 9
Due to changes in react router 9, some of the providers needed to be moved to the root layout.
