+++ src/web/src/index.tsx
@@ -1,45 +1,26 @@
-import React from "react";
+import { StrictMode } from "react";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { Auth0Provider } from "@auth0/auth0-react";
 import { createRoot } from "react-dom/client";
 import { SnackbarProvider } from "notistack";
 import NiceModal from "@ebay/nice-modal-react";
 import { I18nextProvider } from "react-i18next";
 import i18n from "./i18n";
 
-const queryClient = new QueryClient();
-const domain = (import.meta.env.VITE_APP_AUTH0_DOMAIN as string) || "";
-const clientId = (import.meta.env.VITE_APP_AUTH0_CLIENT_ID as string) || "";
-
 const container = document.getElementById("root") as Element;
 const root = createRoot(container);
 
 root.render(
-  <React.StrictMode>
-    <Auth0Provider
-      domain={domain}
-      clientId={clientId}
-      authorizationParams={{
-        audience: "https://api.kdvmanager.nl/",
-        redirect_uri: window.location.origin,
-      }}
-    >
-      <QueryClientProvider client={queryClient}>
-        <I18nextProvider i18n={i18n}>
-          <SnackbarProvider>
-            <NiceModal.Provider>
-              <App />
-            </NiceModal.Provider>
-          </SnackbarProvider>
-        </I18nextProvider>
-        <ReactQueryDevtools initialIsOpen={false} />
-      </QueryClientProvider>
-    </Auth0Provider>
-  </React.StrictMode>,
+  <StrictMode>
+    <I18nextProvider i18n={i18n}>
+      <SnackbarProvider>
+        <NiceModal.Provider>
+          <App />
+        </NiceModal.Provider>
+      </SnackbarProvider>
+    </I18nextProvider>
+  </StrictMode>,
 );
 
 // If you want to start measuring performance in your app, pass a function
